Speeding up empirical prior evaluation by ~5x
The empirical priors evaluation was being done in each individual plot, which
meant that if you had e.g. 5 empirical priors and 5 plots, the empirical priors
would be evaluated 25 times. Now, the empirical priors are evaluated once per
batch in the plot stack, and the weights are passed to the individual plots.